[add] <DB> 1. add table "token", "member_token" 2. delete column member.token_version 3. rename column member.mid -> member_id 4. rename column member.id -> login_id 5. rename column member.pw -> password <Java> 1. refreshTokenの追加のため新しいmodule作成中 2. DB Column名変更のため、リポジトリ更新

+++ src/main/java/com/hanyahunya/gitRemind/member/service/MemberServiceImpl.java
@@ -1,10 +1,11 @@
 package com.hanyahunya.gitRemind.member.service;
 
 import com.hanyahunya.gitRemind.member.dto.*;
-import com.hanyahunya.gitRemind.token.service.TokenService;
+import com.hanyahunya.gitRemind.token.service.AccessTokenService;
 import com.hanyahunya.gitRemind.util.ResponseDto;
 import com.hanyahunya.gitRemind.member.entity.Member;
 import com.hanyahunya.gitRemind.member.repository.MemberRepository;
+import com.hanyahunya.gitRemind.util.service.EncodeService;
 import lombok.RequiredArgsConstructor;
 
 import java.util.Optional;
@@ -13,19 +14,19 @@ import java.util.UUID;
 @RequiredArgsConstructor
 public class MemberServiceImpl implements MemberService {
     private final MemberRepository memberRepository;
-    private final TokenService tokenService;
-    private final PwEncodeService pwEncodeService;
+    private final AccessTokenService accessTokenService;
+    private final EncodeService encodeService;
 
     @Override
     public ResponseDto<JwtResponseDto> join(JoinRequestDto joinRequestDto) {
         UUID uuid = UUID.randomUUID();
         Member member = joinRequestDto.dtoToEntity();
-        member.setMid(uuid.toString());
+        member.setMemberId(uuid.toString());
         // パスワードencode
-        member.setPw(pwEncodeService.encode(member.getPw()));
+        member.setPassword(encodeService.encode(member.getPassword()));
         boolean success = memberRepository.saveMember(member);
         if (success) {
-            String token = tokenService.generateToken(member);
+            String token = accessTokenService.generateToken(member);
             return ResponseDto.success("会員登録成功", JwtResponseDto.set(token));
         } else {
             return ResponseDto.fail("会員登録失敗", null);
@@ -36,10 +37,10 @@ public class MemberServiceImpl implements MemberService {
     public ResponseDto<JwtResponseDto> login(LoginRequestDto loginRequestDto) {
         Optional<Member> optionalMember = memberRepository.validateMember(loginRequestDto.dtoToEntity());
         if (optionalMember.isPresent()) {
-            String reqPw = loginRequestDto.getPw();
+            String reqPw = loginRequestDto.getPassword();
             Member member = optionalMember.get();
-            if (pwEncodeService.matches(reqPw, member.getPw())) {
-                return ResponseDto.success("ログイン成功", JwtResponseDto.set(tokenService.generateToken(member)));
+            if (encodeService.matches(reqPw, member.getPassword())) {
+                return ResponseDto.success("ログイン成功", JwtResponseDto.set(accessTokenService.generateToken(member)));
             } else {
                 return ResponseDto.fail("ログイン失敗");
             }
@@ -49,8 +50,8 @@ public class MemberServiceImpl implements MemberService {
     }
 
     @Override
-    public ResponseDto<MemberInfoResponseDto> getInfo(String mid) {
-        return memberRepository.findMemberByMid(mid)
+    public ResponseDto<MemberInfoResponseDto> getInfo(String memberId) {
+        return memberRepository.findMemberByMemberId(memberId)
                 .map(member ->
                         ResponseDto.success("ユーザー情報ロード成功", MemberInfoResponseDto.set(member.getEmail()))
                 )
@@ -59,12 +60,16 @@ public class MemberServiceImpl implements MemberService {
 
     @Override
     public ResponseDto<Void> deleteMember(DeleteMemberRequestDto requestDto) {
-        requestDto.setPw(pwEncodeService.encode(requestDto.getPw()));
-        if(memberRepository.deleteMember(requestDto.toEntity())) {
-            return ResponseDto.success("ユーザー退会成功");
-        } else {
-            return ResponseDto.success("ユーザー退会失敗");
+        Optional<Member> optionalMember = memberRepository.validateMember(requestDto.toEntity());
+        if (optionalMember.isPresent()) {
+            Member dbMember = optionalMember.get();
+            if (encodeService.matches(requestDto.getPassword(), dbMember.getPassword())) {
+                if(memberRepository.deleteMember(requestDto.toEntity())) {
+                    return ResponseDto.success("ユーザー退会成功");
+                }
+            }
         }
+        return ResponseDto.fail("ユーザー退会失敗");
     }
 
     @Override
@@ -72,7 +77,8 @@ public class MemberServiceImpl implements MemberService {
         Optional<Member> optionalMember = memberRepository.validateMember(requestDto.toEntity());
         if (optionalMember.isPresent()) {
             Member dbMember = optionalMember.get();
-            if (pwEncodeService.matches(requestDto.getPw(), dbMember.getPw())) {
+            if (encodeService.matches(requestDto.getPassword(), dbMember.getPassword())) {
+                requestDto.setPassword(null);
                 if (memberRepository.updateMember(requestDto.toEntity())) {
                     return ResponseDto.success("ユーザー情報更新成功");
                 }
